fix: satisfy pkg-file-count limit for factorysessionexecution
